Valid Palindrome

+++ Valid Palindrome.cpp
@@ -28,7 +28,8 @@ public:
 class Solution {
 public:
     bool isPalindrome(string s) {
-        int l = 0, r = s.size()-1;
+        //int l = 0, r = s.size()-1;
+        int l = 0, r = (int)s.size()-1;
         while(l < r) {
             if(!isalnum(s[l])) l++;
             else if(!isalnum(s[r])) r--;
